+++ History.md
@@ -105,6 +105,9 @@
 * Add login hooks on the client: `Accounts.onLogin` and
   `Accounts.onLoginFailure`. #3572
 
+* Fix crash in `meteor publish` when the package is inside an app which uses
+  Cordova. #3676
+
 * Upgraded dependencies:
 
   - node: 0.10.36 (from 0.10.33)